+++ pkg/gardenlet/bootstrappers/seed_config.go
